refactor: densify Slice D allowlisted app test suites under 400 lines
Extract shared fixtures for event-handler, civilian, production, region-map,
and province-overlay hosts so AC10 leftover families shrink without new partN shards.
