+++ app/test/ct_region_map_widget_part2_test.dart
@@ -4,17 +4,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:colonizethis_map/colonizethis_map.dart';
-import 'package:colonizethis_models/colonizethis_models.dart'
-    show
-        OpenCivilianUnitsPanelEvent,
-        OpenNavalMissionMenuEvent;
-
 import 'package:colonizethis_app/features/game/flame/caches/resource_icon_cache.dart';
 import 'package:colonizethis_app/features/game/flame/region_map/region_map.dart'
-    show BaseLayerDisplayMode, CtMapVisibilityMode;
+    show CtMapVisibilityMode;
 import 'package:colonizethis_app/features/game/flame/tilesets/tilesets.dart';
 
 import 'ct_region_map_test_support.dart';
+import 'ct_region_map_widget_part2_support.dart';
 
 Finder get _map => ctRegionMapFinder();
 
@@ -170,45 +166,12 @@ void main() {
       'tap on civilian marker tile invokes civilian callback and suppresses detail tap callback',
       (WidgetTester tester) async {
         const markerTileKey = 'oldWorld|pMarker|0|0';
-        final region = ctRegionMapMiniLandStrip(
-          base: ctRegionMapTestOldWorldRegion(),
-          width: 1,
-          height: 1,
-          cellSize: 24,
-          regionCellId: 'pMarker',
-          displayName: 'Marker Province',
-          civilianTileMarkers: [
-            ctRegionMapCivilianMarker(
-              tileKey: markerTileKey,
-              x: 0,
-              y: 0,
-              localProvinceId: 'pMarker',
-            ),
-          ],
-        );
-        String? tappedCivilianTileKey;
-        String? detailTileKey;
-        String? selectedProvinceId;
-        final (bus, openedPanels) = ctRegionMapBusCapture<OpenCivilianUnitsPanelEvent>();
-        await pumpCtRegionMapTest(
-          tester,
-          region: region,
-          width: 64,
-          height: 64,
-          cellSizePx: 32,
-          bus: bus,
-          onCivilianTileStateChanged: (tileKey) =>
-              tappedCivilianTileKey = tileKey,
-          onMapTileTappedForDetail: (tileKey) => detailTileKey = tileKey,
-          onProvinceSelected: (id) => selectedProvinceId = id,
-        );
-        await tapCtRegionMap(tester);
+        final (tappedCivilianTileKey, openedPanels) =
+            await pumpAndTapCivilianMarker(tester);
         expect(tappedCivilianTileKey, equals(markerTileKey));
         expect(openedPanels, hasLength(1));
         expect(openedPanels.single.tileScopeTileKey, equals(markerTileKey));
         expect(openedPanels.single.initialSelectedUnitId, equals('u_builder'));
-        expect(detailTileKey, isNull);
-        expect(selectedProvinceId, isNull);
       },
       timeout: const Timeout(Duration(seconds: 5)),
     );
@@ -217,33 +180,7 @@ void main() {
       'tapping fleet marker emits naval mission menu event',
       (WidgetTester tester) async {
         const markerTileKey = 'oldWorld|sMarker|0|0';
-        final region = ctRegionMapMiniLandStrip(
-          base: ctRegionMapTestOldWorldRegion(),
-          width: 1,
-          height: 1,
-          cellSize: 24,
-          regionCellId: 'sMarker',
-          displayName: 'Marker Sea',
-          sea: true,
-          fleetTileMarkers: [
-            ctRegionMapFleetMarker(
-              tileKey: markerTileKey,
-              x: 0,
-              y: 0,
-              locationScopeKey: 'sea:oldWorld|fleet_scope',
-            ),
-          ],
-        );
-        final (bus, openedMenus) = ctRegionMapBusCapture<OpenNavalMissionMenuEvent>();
-        await pumpCtRegionMapTest(
-          tester,
-          region: region,
-          width: 64,
-          height: 64,
-          cellSizePx: 32,
-          bus: bus,
-        );
-        await tapCtRegionMap(tester);
+        final openedMenus = await pumpAndTapFleetMarker(tester);
         expect(openedMenus, hasLength(1));
         expect(
           openedMenus.single.locationScopeKey,
@@ -302,38 +239,7 @@ void main() {
     testWidgets(
       'tap on a town tile still invokes map tile and province selection callbacks',
       (WidgetTester tester) async {
-        final region = ctRegionMapMiniLandStrip(
-          base: ctRegionMapTestOldWorldRegion(),
-          width: 1,
-          height: 1,
-          cellSize: 24,
-          regionCellId: 'pTown',
-          displayName: 'Town Province',
-          townMarkers: const [
-            TownMarkerView(
-              x: 0,
-              y: 0,
-              provinceId: 'pTown',
-              isCoastal: false,
-              isPort: false,
-              touchesSea: false,
-              townDevelopmentLevel: 1,
-              townIconStyle: 'euro',
-            ),
-          ],
-        );
-        String? selectedId;
-        String? detailTileKey;
-        await pumpCtRegionMapTest(
-          tester,
-          region: region,
-          onProvinceSelected: (id) => selectedId = id,
-          onMapTileTappedForDetail: (tk) => detailTileKey = tk,
-          width: 64,
-          height: 64,
-          cellSizePx: 32,
-        );
-        await tapCtRegionMap(tester);
+        final (selectedId, detailTileKey) = await pumpAndTapTownMarker(tester);
         expect(selectedId, equals('oldWorld|pTown'));
         expect(detailTileKey, equals('oldWorld|pTown|0|0'));
       },
@@ -415,16 +321,7 @@ void main() {
       'map renders with resource icons across resource base-layer modes '
       '(SPEC/ui/map-widget.md § Base layer display mode)',
       (WidgetTester tester) async {
-        await tester.runAsync(preloadCtRegionMapRoadAssets);
-        final region = ctRegionMapTestOldWorldRegion();
-        for (final mode in [
-          BaseLayerDisplayMode.terrainAndResources,
-          BaseLayerDisplayMode.terrainAndResourcesImprovementLabels,
-          BaseLayerDisplayMode.terrainAndResourcesImprovementsRoads,
-        ]) {
-          await pumpCtRegionMapTest(tester, region: region, baseLayerDisplayMode: mode);
-          expect(_map, findsOneWidget);
-        }
+        await pumpCtRegionMapResourceBaseModes(tester);
       },
       timeout: const Timeout(Duration(seconds: 15)),
     );
@@ -432,18 +329,7 @@ void main() {
     testWidgets(
       'roads mode renders for non-64 cell sizes with transport overlay assets preloaded',
       (WidgetTester tester) async {
-        await tester.runAsync(preloadCtRegionMapRoadAssets);
-        final region = ctRegionMapTestOldWorldRegion();
-        for (final cellSize in [16.0, 32.0, 96.0]) {
-          await pumpCtRegionMapTest(
-            tester,
-            region: region,
-            cellSizePx: cellSize,
-            baseLayerDisplayMode:
-                BaseLayerDisplayMode.terrainAndResourcesImprovementsRoads,
-          );
-          expect(_map, findsOneWidget);
-        }
+        await pumpCtRegionMapRoadsCellSizes(tester);
       },
       timeout: const Timeout(Duration(seconds: 12)),
     );
