+++ app/test/province_detail_overlay_host_support_test.dart
@@ -1,100 +1,13 @@
 // Unit pins for the shared province-detail overlay host support extracted in
 // Refs #3594 (work item 7 — resolve flame-host ↔ widget duplication/coupling).
-import 'package:colonizethis_app/core/services/game_service/game_service.dart'
-    show GameMapData;
-import 'package:colonizethis_app/features/game/flame/caches/per_player_work_target_selection_cache.dart';
 import 'package:colonizethis_app/features/game/flame/overlays/province_detail_overlay_host_support.dart';
-import 'package:colonizethis_data/colonizethis_data.dart'
-    show
-        MapTopology,
-        Resource,
-        TileMapResult,
-        TopologyNode,
-        TopologyNodeType,
-        kTechIdMoldboardPlow;
 import 'package:colonizethis_logic/colonizethis_logic.dart'
-    show PlayerView, buildPlayerView, kWorkTargetBuildImprovement;
-import 'package:colonizethis_map/colonizethis_map.dart';
+    show kWorkTargetBuildImprovement;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter_test/flutter_test.dart';
 
-const String _kPlayerId = 'gp1';
-const String _kTileKey = 'oldWorld|p1|0|0';
-
-Game _minimalGame() => Game(
-      id: 'g_support',
-      worldState: const WorldState(
-        turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-        oldWorld: RegionData(provinces: [], units: []),
-        newWorld: RegionData(provinces: [], units: []),
-      ),
-      players: const [
-        Player(
-          id: _kPlayerId,
-          displayName: 'Human',
-          isHuman: true,
-          capitalProvinceId: '',
-        ),
-      ],
-      minorNations: const [],
-      tribes: const [],
-    );
-
-RegionMapViewData _emptyRegion() => const RegionMapViewData(
-      regionId: 'oldWorld',
-      width: 1,
-      height: 1,
-      cellSize: 16,
-      cells: [],
-      capitalMarkers: [],
-      portMarkers: [],
-      factionColors: {},
-      greatPowerFactionIds: {},
-      terrainColors: {},
-      provincePoliticalOwnerByPrefixedProvinceId: {},
-    );
-
-PlayerView _playerView(Game game) =>
-    buildPlayerView(game, const MapTopology(nodes: [], edges: []), _kPlayerId);
-
-ProvinceDetailShortcutCallbacks _callbacks({
-  required Game game,
-  required String? selectedTileKey,
-  required bool exploreEnabled,
-  required bool prospectEnabled,
-  required bool buildImprovementEnabled,
-  required bool buildRoadEnabled,
-  required bool buildFortEnabled,
-  required bool buildPortEnabled,
-  required bool purchaseLandEnabled,
-  bool upgradeTownEnabled = false,
-  String? upgradeTownTargetTileKey,
-  String provinceId = 'oldWorld|p1',
-  required AppEventBus bus,
-}) =>
-    buildProvinceDetailShortcutCallbacks(
-      game: game,
-      humanPlayerId: _kPlayerId,
-      region: _emptyRegion(),
-      playerView: _playerView(game),
-      workTargetSelectionCache:
-          PerPlayerWorkTargetSelectionCache(strategies: const {}),
-      draftOrders: const Orders(),
-      mapData: null,
-      selectedTileKey: selectedTileKey,
-      exploreEnabled: exploreEnabled,
-      prospectEnabled: prospectEnabled,
-      buildImprovementEnabled: buildImprovementEnabled,
-      buildRoadEnabled: buildRoadEnabled,
-      buildFortEnabled: buildFortEnabled,
-      buildPortEnabled: buildPortEnabled,
-      purchaseLandEnabled: purchaseLandEnabled,
-      provinceId: provinceId,
-      upgradeTownEnabled: upgradeTownEnabled,
-      upgradeTownTargetTileKey: upgradeTownTargetTileKey,
-      bus: bus,
-    );
+import 'province_detail_overlay_host_support_fixtures.dart';
 
 void main() {
   suppressLogsForTests();
@@ -102,14 +15,20 @@ void main() {
   group('resolveProvinceDetailDisplayId', () {
     test('returns empty string for a null tile key', () {
       expect(
-        resolveProvinceDetailDisplayId(region: _emptyRegion(), tileKey: null),
+        resolveProvinceDetailDisplayId(
+          region: provinceDetailEmptyRegion(),
+          tileKey: null,
+        ),
         isEmpty,
       );
     });
 
     test('returns empty string for an empty tile key', () {
       expect(
-        resolveProvinceDetailDisplayId(region: _emptyRegion(), tileKey: ''),
+        resolveProvinceDetailDisplayId(
+          region: provinceDetailEmptyRegion(),
+          tileKey: '',
+        ),
         isEmpty,
       );
     });
@@ -127,8 +46,8 @@ void main() {
     });
 
     test('returns all-null callbacks when no tile is selected', () {
-      final callbacks = _callbacks(
-        game: _minimalGame(),
+      final callbacks = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
         selectedTileKey: null,
         exploreEnabled: true,
         prospectEnabled: true,
@@ -148,9 +67,9 @@ void main() {
     });
 
     test('returns all-null callbacks when every action is disabled', () {
-      final callbacks = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final callbacks = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -169,9 +88,9 @@ void main() {
     });
 
     test('exposes only the enabled action callback (per-action gating)', () {
-      final exploreOnly = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final exploreOnly = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: true,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -187,9 +106,9 @@ void main() {
       expect(exploreOnly.onBuildRoadTap, isNull);
       expect(exploreOnly.onPurchaseLandTap, isNull);
 
-      final prospectOnly = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final prospectOnly = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: false,
         prospectEnabled: true,
         buildImprovementEnabled: false,
@@ -205,9 +124,9 @@ void main() {
       expect(prospectOnly.onBuildRoadTap, isNull);
       expect(prospectOnly.onPurchaseLandTap, isNull);
 
-      final buildOnly = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final buildOnly = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: true,
@@ -223,9 +142,9 @@ void main() {
       expect(buildOnly.onBuildRoadTap, isNull);
       expect(buildOnly.onPurchaseLandTap, isNull);
 
-      final buildRoadOnly = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final buildRoadOnly = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -241,9 +160,9 @@ void main() {
       expect(buildRoadOnly.onBuildRoadTap, isNotNull);
       expect(buildRoadOnly.onPurchaseLandTap, isNull);
 
-      final purchaseLandOnly = _callbacks(
-        game: _minimalGame(),
-        selectedTileKey: _kTileKey,
+      final purchaseLandOnly = provinceDetailCallbacks(
+        game: provinceDetailMinimalGame(),
+        selectedTileKey: provinceDetailSupportTileKey,
         exploreEnabled: false,
         prospectEnabled: false,
         buildImprovementEnabled: false,
@@ -262,86 +181,15 @@ void main() {
   });
 
   group('provinceExtractionSnapshotPreview projection (Refs #4064)', () {
-    const provinceId = 'oldWorld|p1';
-    const tk = 'oldWorld|p1|0|0';
-
-    Game gameWithImprovedGrain({required String ownerId}) {
-      return Game(
-        id: 'g_extraction_preview',
-        capitalTileGrainBonusPerTurn: 0,
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 2),
-          oldWorld: RegionData(
-            provinces: [
-              Province(
-                id: provinceId,
-                regionId: 'oldWorld',
-                ownerId: ownerId,
-                townDevelopmentLevel: 4,
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-          tileState: TileMapState().setImprovement(tk, 2).setRoadLevel(tk, 4),
-          resourceByTileKey: const {tk: 'grain'},
-          tileKeysByRegionAndProvince: const {
-            'oldWorld': {
-              'oldWorld|p1': [tk],
-            },
-          },
-        ),
-        players: [
-          Player(
-            id: ownerId,
-            displayName: 'GP',
-            isHuman: true,
-            capitalProvinceId: provinceId,
-            capitalTile: const CapitalTile(
-              regionId: 'oldWorld',
-              provinceId: 'oldWorld|p1',
-              x: 0,
-              y: 0,
-            ),
-            techUnlocked: const {kTechIdMoldboardPlow: true},
-          ),
-        ],
-      );
-    }
-
-    GameMapData mapDataForProjection() {
-      final tileMap = TileMapResult(
-        width: 1,
-        height: 1,
-        grid: const [
-          ['p1'],
-        ],
-        resourceGrid: const [
-          [Resource.grain],
-        ],
-      );
-      return (
-        combinedTopology: const MapTopology(
-          nodes: [
-            TopologyNode(
-              id: 'p1',
-              regionId: 'oldWorld',
-              type: TopologyNodeType.province,
-            ),
-          ],
-          edges: [],
-        ),
-        tileMapByRegion: {'oldWorld': tileMap},
-        topologyByRegion: const <String, MapTopology>{},
-        warpLinks: null,
-      );
-    }
+    const provinceId = provinceDetailSupportProvinceId;
+    const tk = provinceDetailSupportTileKey;
 
     test('projects non-empty Extraction without last-turn history', () {
-      final game = gameWithImprovedGrain(ownerId: 'gp1');
+      final game = provinceDetailGameWithImprovedGrain(ownerId: 'gp1');
       final snap = provinceExtractionSnapshotPreview(
         game: game,
         provinceId: provinceId,
-        mapData: mapDataForProjection(),
+        mapData: provinceDetailMapDataForProjection(),
       );
       expect(snap, isNotNull);
       expect(snap!.ownerId, 'gp1');
@@ -351,9 +199,9 @@ void main() {
     test(
       'ownership change: preview attributes to new owner without Extraction write',
       () {
-        final mapData = mapDataForProjection();
+        final mapData = provinceDetailMapDataForProjection();
         final before = provinceExtractionSnapshotPreview(
-          game: gameWithImprovedGrain(ownerId: 'gp1'),
+          game: provinceDetailGameWithImprovedGrain(ownerId: 'gp1'),
           provinceId: provinceId,
           mapData: mapData,
         );
@@ -362,7 +210,7 @@ void main() {
         expect(before.byCommodity['grain']!.full, greaterThan(0));
 
         final after = provinceExtractionSnapshotPreview(
-          game: gameWithImprovedGrain(ownerId: 'gp2'),
+          game: provinceDetailGameWithImprovedGrain(ownerId: 'gp2'),
           provinceId: provinceId,
           mapData: mapData,
         );
@@ -374,7 +222,7 @@ void main() {
     );
 
     test('returns null when map data is missing', () {
-      final game = gameWithImprovedGrain(ownerId: 'gp1');
+      final game = provinceDetailGameWithImprovedGrain(ownerId: 'gp1');
       expect(
         provinceExtractionSnapshotPreview(
           game: game,
@@ -391,48 +239,8 @@ void main() {
       () {
         // Host preview has no draftOrders parameter; staged WorkOrders must not
         // invent yields until turn resolution updates Game.tileState.
-        final unresolved = Game(
-          id: 'g_extraction_draft_ignore',
-          capitalTileGrainBonusPerTurn: 0,
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 2),
-            oldWorld: RegionData(
-              provinces: [
-                Province(
-                  id: provinceId,
-                  regionId: 'oldWorld',
-                  ownerId: 'gp1',
-                  townDevelopmentLevel: 4,
-                ),
-              ],
-            ),
-            newWorld: const RegionData(),
-            // No improvement yet — draft build_improvement would raise this.
-            tileState: const TileMapState(),
-            resourceByTileKey: const {tk: 'grain'},
-            tileKeysByRegionAndProvince: const {
-              'oldWorld': {
-                'oldWorld|p1': [tk],
-              },
-            },
-          ),
-          players: [
-            Player(
-              id: 'gp1',
-              displayName: 'GP',
-              isHuman: true,
-              capitalProvinceId: provinceId,
-              capitalTile: const CapitalTile(
-                regionId: 'oldWorld',
-                provinceId: 'oldWorld|p1',
-                x: 0,
-                y: 0,
-              ),
-              techUnlocked: const {kTechIdMoldboardPlow: true},
-            ),
-          ],
-        );
-        final mapData = mapDataForProjection();
+        final unresolved = provinceDetailUnresolvedExtractionGame();
+        final mapData = provinceDetailMapDataForProjection();
         final beforeDraft = provinceExtractionSnapshotPreview(
           game: unresolved,
           provinceId: provinceId,
@@ -440,7 +248,6 @@ void main() {
         );
         expect(beforeDraft?.byCommodity['grain'], isNull);
 
-        // Local draft Orders exist mid-turn but are never passed into preview.
         final draftOrders = Orders(
           workOrdersByPlayerId: {
             'gp1': [
@@ -461,7 +268,7 @@ void main() {
         );
         expect(afterDraftPresence, beforeDraft);
 
-        final resolved = gameWithImprovedGrain(ownerId: 'gp1');
+        final resolved = provinceDetailGameWithImprovedGrain(ownerId: 'gp1');
         final afterResolution = provinceExtractionSnapshotPreview(
           game: resolved,
           provinceId: provinceId,
